[plugin] handle unknown exceptions when checking device ids
Although get_server() is documented to return NotFoundException
which is an alias for ResourceNotFound [0], it does report different
errors at times.

(Recently ResourceNotFound has been aliased as NotFoundException [1]).

In a user environment, it's returned ValueError. While it's been fixed
in OpenStack [2], I don't see any why we need to differentiate the
different possible exceptions here. In this plugin, we try to classify
a port-down as 'warning' if we can determine their instances were
shutdown; otherwise classify them as 'critical'.
If, for whatever reason, we can't detertmine the reason for a port
being down, the status defaults to 'critical'.

[0] https://docs.openstack.org/openstacksdk/latest/user/proxies/compute.html
[1] https://github.com/openstack/openstacksdk/commit/14c1ebbeb8e6d7f89bd96476e6fca8ce822aebad
[2] https://github.com/openstack/openstacksdk/commit/fd2c41c69715ba39588c18224aef4e8734ed3ce1

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>