Restructure and run func tests locally (#197)
Restructure the charm so it works with our latest templates, and work on the func tests so they can run locally on the runner without an external juju controller.

* Pin pbr to fix install

pbr release 6.1.1 added setuptools >= 64.0.0 as a build dependency.
However, due to other constraints the maximum version of
setuptools available to install is 59.6.0.



Co-authored-by: Gabriel Cocenza <gabriel.cocenza@canonical.com>